requirements.txt
Django==5.0.4
PyMuPDF==1.24.2
databricks-sdk==0.26.0
databricks-vectorsearch==0.22
langchain==0.1.12
llama-index==0.10.16
mlflow==2.10.1
pandas==1.5.3
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg==3.1.18
pydantic==1.10.9
pyspark==3.5.1
python-dotenv==1.0.1
pytz==2023.4
transformers==4.36.0
